+++ packages/lumx-vue/src/utils/ClickAway/index.ts
@@ -1,0 +1,2 @@
+export { ClickAwayProvider } from './ClickAwayProvider';
+export { useClickAway } from './useClickAway';